fix: add a data type from int to float

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 monthly_income = input("Enter your monthly income: ")
 monthly_expenses = input("Enter your monthly expenses: ")
 
-monthly_saving = int(monthly_income) - int(monthly_expenses)
+monthly_saving = float(monthly_income) - float(monthly_expenses)
 
 
 projected_saving = monthly_saving * 12 + (monthly_saving * 12 * 0.05)
